requirements_production.txt
# Core
Flask==3.0.0
Flask-SocketIO==5.3.5
python-dotenv==1.0.0

# Database
mysql-connector-python==8.2.0
pymysql==1.1.0

# Security
bcrypt==4.1.2
cryptography==41.0.7

# Face Recognition & AI
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
dlib==19.24.2

# Image Processing
Pillow==10.1.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Notifications
twilio==8.11.0

# Export
reportlab==4.0.7
matplotlib==3.8.2

# Monitoring
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Utils
requests==2.31.0
python-dateutil==2.8.2
qrcode==7.4.2
